Optimize MainGui import in GuiUtility
The MainGui import in 'gui_utility.py' has been optimized to improve application performance. It is now imported only during the type checking phase of 'mypy' and is no longer present when the application is running. The reference to MainGui in the GuiUtility class constructor has also been adjusted to conform with this modification.

Signed-off-by: Jose Borreguero <borreguero@gmail.com>