Added expansion checks to unit test in TableData class
Additional checks have been integrated into the test/unit/RefRed/test_tabledata.py script. These ensure not only that the size of the table is preserved when rows are removed but also that the last items in the table are None after removal. These improvements enhance the robustness and accuracy of the unit tests.

Signed-off-by: Jose Borreguero <borreguero@gmail.com>